--- src/components/style-Message.scss
@@ -1,12 +1,0 @@
-.firstBlock{ 
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-.article{
-    color: blue;
-}
-.myName{
-    color: green;
-}
